+++ ios/Classes/SolanaWalletAdapterPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface SolanaWalletAdapterPlugin : NSObject<FlutterPlugin>
-@end